+++ CHANGELOG.md
@@ -55,6 +55,8 @@
 ### Version 5.1.0-alpha2
 
 - Added support for full-screen functionality in Android Exoplayer [#1730](https://github.com/react-native-community/react-native-video/pull/1730)
+### next
+* Added PictureInPicture support for AndroidExoplayer API >= 28. [#1776](https://github.com/react-native-community/react-native-video/pull/1776)
 
 ### Version 5.1.0-alpha1